+++ routes/question-your-situation-no-income/schema.js
@@ -5,11 +5,12 @@ const Schema = {
     isIn: {
       errorMessage: 'errors.multipleChoiceGeneric',
       options: [[
-        'lost-job-employer-closed','self-employed-closed', 
-        'unpaid-leave-to-care', 
+        'lost-job-employer-closed','self-employed-closed',
+        'unpaid-leave-to-care',
         'sick-or-quarantined', 'parental-recently-cant-return',
-        'student_2019_20', 
+        'student_2019_20',
         'ei-recently-claim-ended',
+        'none-of-the-above',
       ]],
     },
   },