+++ routes/question-your-situation-unchanged-income/schema.js
@@ -4,7 +4,16 @@ const Schema = {
   unchanged_income: {
     isIn: {
       errorMessage: 'errors.multipleChoiceGeneric',
-      options: [['wfh', 'paid-leave', 'retired', 'student_2019_20', 'high_school_grad']],
+      options: [
+        [
+          'wfh',
+          'paid-leave',
+          'retired',
+          'student_2019_20',
+          'high_school_grad',
+          'none-of-the-above',
+        ],
+      ],
     },
   },
 }