+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 All notable changes to this product will be documented in this file.
 
+## 2020-05-28
+
+### Changed:
+* Added none of the above to *No Income*, *Some Income*, and *Unchanged Income* questions.
+
 ## 2020-05-27
 
-## Removed:
+### Removed:
 * Removed the ei_regular_cerb and ei_sickness_cerb cards and updated the calculations/tests
 
 ## 2020-05-22